Migrate build/ to build/kernel (step 1)
This is the step 1 to move the local directory for kernel/build
from build/ to build/kernel.

Benefits of this include:
- Not conflicting with platform/build
- Attempt to join an project that massively reduce presubmit /
  postsubmit latency times, which requires an unnested repo
- Aesthetic

In Step 1, symlinks are created so this patch does not affect downstream
trees. Subsequent CLs will be submitted to remove these symlinks once
the code is able to handle both build/ and build/kernel.

Test: BUILD_CONFIG=... build/build.sh
Bug: 204425264
Change-Id: I793e2814bbcdcdaa921ae5979457401ff439c158