Align default revision with Android version
Certain projects may move forward on their master branch breaking
compatibility with other Android-version-specific projects. This happend
e.g. with the 'kernel/build' project where on the master branch certain
path' and locations have changed.

Thus we have to pin the default revision to the Android and kernel
version we want to build for. This also fixes issue #7.

Change-Id: Icf22094caf3521d506acaeb9a3c3f9265d1a877f